# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API do sistema GuardPoint (controle de rondas/postos).",
        "title": "GuardPoint API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/postos": {
            "get": {
                "tags": [
                    "postos"
                ],
                "summary": "Lista os postos da empresa",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filtra somente postos ativos (true/false)",
                        "name": "ativos",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.Posto"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "postos"
                ],
                "summary": "Cria um posto",
                "parameters": [
                    {
                        "description": "Dados do posto",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.CreatePostoRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.Posto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/postos/{id}": {
            "get": {
                "tags": [
                    "postos"
                ],
                "summary": "Busca um posto pelo ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do posto (uuid)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.Posto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "postos"
                ],
                "summary": "Atualiza um posto",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do posto (uuid)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Campos a atualizar",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.UpdatePostoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.Posto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "postos"
                ],
                "summary": "Desativa um posto",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do posto (uuid)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/postos/{id}/supervisores": {
            "get": {
                "description": "Retorna a lista de IDs dos supervisores que estao vinculados a este posto e portanto recebem alertas do mesmo.",
                "tags": [
                    "postos"
                ],
                "summary": "Lista os supervisores vinculados a um posto (admin/supervisor)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do posto (uuid)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.PostoSupervisorResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "posto_id invalido",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "erro interno",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Um supervisor vinculado a um posto recebera alertas apenas desse posto (alem de precisar estar na configuracao de escalonamento). Admins recebem alertas de todos os postos independente de vinculo.",
                "tags": [
                    "postos"
                ],
                "summary": "Vincula um supervisor a um posto (somente admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do posto (uuid)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ID do supervisor a ser vinculado",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.PostoSupervisorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.PostoSupervisorResponse"
                        }
                    },
                    "400": {
                        "description": "posto_id ou supervisor_id invalido",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "posto ou supervisor nao encontrado",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "erro interno",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/postos/{id}/supervisores/{supervisorId}": {
            "delete": {
                "description": "Remove o supervisor do posto. Ele deixara de receber alertas especificos deste posto (a menos que seja admin, que continua recebendo todos).",
                "tags": [
                    "postos"
                ],
                "summary": "Remove o vinculo de um supervisor com um posto (somente admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do posto (uuid)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID do supervisor (uuid)",
                        "name": "supervisorId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "posto_id ou supervisor_id invalido",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "vinculo nao encontrado",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "erro interno",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/usuarios/{supervisorId}/postos": {
            "get": {
                "description": "Retorna a lista de postos (com nome) aos quais o supervisor esta vinculado. Util para a UI mostrar/ocultar alertas por posto.",
                "tags": [
                    "postos"
                ],
                "summary": "Lista os postos vinculados a um supervisor (admin/supervisor)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do supervisor (uuid)",
                        "name": "supervisorId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.SupervisorPostoResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "supervisor_id invalido",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "erro interno",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_guardpoint_guardpoint-server_internal_model.CreatePostoRequest": {
            "type": "object",
            "required": [
                "latitude",
                "longitude",
                "nome"
            ],
            "properties": {
                "latitude": {
                    "type": "number",
                    "example": -23.55052
                },
                "longitude": {
                    "type": "number",
                    "example": -46.633308
                },
                "nome": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2,
                    "example": "Portaria Principal"
                },
                "raio_m": {
                    "type": "integer",
                    "maximum": 5000,
                    "minimum": 10,
                    "example": 100
                }
            }
        },
...
        "github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "mensagem descritiva do erro"
                }
            }
        },
...
        "github_com_guardpoint_guardpoint-server_internal_model.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "operacao realizada com sucesso"
                }
            }
        },
        "github_com_guardpoint_guardpoint-server_internal_model.Posto": {
            "type": "object",
            "properties": {
                "ativo": {
                    "type": "boolean",
                    "example": true
                },
                "created_at": {
                    "type": "string"
                },
                "empresa_id": {
                    "type": "string",
                    "example": "660e8400-e29b-41d4-a716-446655440000"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "latitude": {
                    "type": "number",
                    "example": -23.55052
                },
                "longitude": {
                    "type": "number",
                    "example": -46.633308
                },
                "nome": {
                    "type": "string",
                    "example": "Portaria Principal"
                },
                "raio_m": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
        "github_com_guardpoint_guardpoint-server_internal_model.PostoSupervisorRequest": {
            "type": "object",
            "required": [
                "supervisor_id"
            ],
            "properties": {
                "supervisor_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "github_com_guardpoint_guardpoint-server_internal_model.PostoSupervisorResponse": {
            "type": "object",
            "properties": {
                "posto_id": {
                    "type": "string",
                    "example": "660e8400-e29b-41d4-a716-446655440000"
                },
                "supervisor_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "github_com_guardpoint_guardpoint-server_internal_model.SupervisorPostoResponse": {
            "type": "object",
            "properties": {
                "posto_id": {
                    "type": "string",
                    "example": "660e8400-e29b-41d4-a716-446655440000"
                },
                "posto_nome": {
                    "type": "string",
                    "example": "Portaria Principal"
                }
            }
        },
...
        "github_com_guardpoint_guardpoint-server_internal_model.UpdatePostoRequest": {
            "type": "object",
            "properties": {
                "ativo": {
                    "type": "boolean"
                },
                "latitude": {
                    "type": "number",
                    "example": -23.551
                },
                "longitude": {
                    "type": "number",
                    "example": -46.634
                },
                "nome": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2,
                    "example": "Portaria Norte"
                },
                "raio_m": {
                    "type": "integer",
                    "maximum": 5000,
                    "minimum": 10,
                    "example": 150
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ]
}